% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Leaf Litter Calc.R, R/Leaf_Litter_Calculation.R
\name{leaf_diff_y2017_2018}
\alias{leaf_diff_y2017_2018}
\title{Calculates the Leaf litter difference across years of collected leaflitter weights(g)}
\usage{
leaf_diff_y2017_2018(
  data = Leaf_litter_filtered,
  Y2017 = Leaf_litter_filtered$Y2017,
  Y2018 = Leaf_litter_filtered$Y2018
)

leaf_diff_y2017_2018(
  data = Leaf_litter_filtered,
  Y2017 = Leaf_litter_filtered$Y2017,
  Y2018 = Leaf_litter_filtered$Y2018
)
}
\arguments{
\item{Y2017}{year 2017 leaf litter weights (numeric, no default)}

\item{Y2018}{year 2018 leaf litter weights (numeric, no default)}

\item{Leaf_litter_filtered}{name of data table (tibble, no default)}

\item{Y2019}{year 2019 leaf litter weights (numeric, no default)}
}
\value{
difference between yearly leaf litter weights (numeric)

difference between yearly leaf litter weights (numeric)
}
\description{
Calculates the Leaf litter difference across years of collected leaflitter weights(g)

Calculates the Leaf litter difference across years of collected leaflitter weights(g)
}
