% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Liner_Model.R
\name{plot_lm}
\alias{plot_lm}
\title{Makes a linear model comparing average diameters and leaf litter weights across the same sites over a number of years}
\usage{
plot_lm(leaf_litter_data)
}
\arguments{
\item{Leaf_liter_and_diameter_combined_avg}{name of data set (tibble, no default)}

\item{diameter_2017_avg}{avg diameter of two species of trees per site (numeric, no default)}

\item{Y2017}{avg leaf litter weight across two species and multiple sites (numeric, no default)}
}
\description{
Makes a linear model comparing average diameters and leaf litter weights across the same sites over a number of years
}
