% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square calc.R
\name{square_calc}
\alias{square_calc}
\title{calculates the area of a square}
\usage{
square_calc(Side_length)
}
\arguments{
\item{Side_length}{length of a side of the square}
}
\value{
area_of_square
}
\description{
calculates the area of a square
}
