% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diameter_Difference_Calculation.R
\name{diameter_diff_y2017_2018}
\alias{diameter_diff_y2017_2018}
\title{Calculates diameter difference from tree species diameters and finds difference between years}
\usage{
diameter_diff_y2017_2018(
  data = Diameters,
  diameter_2017 = Diameters$diameter_2017,
  diameter_2018 = Diameters$diameter_2018
)
}
\arguments{
\item{diameter_2017}{yearly diameter measurement for 2017 (numeric, no default)}

\item{diameter_2018}{yearly diameter measurement for 2018 (numeric, no default)}

\item{Diameters}{name of data table used to calculate diameter difference}
}
\value{
differences between yearly diameters (data frame)
}
\description{
Calculates diameter difference from tree species diameters and finds difference between years
}
